Move gslworkspace and gsl_error_handler_printout() to ratecoeff.h
